02-weaviate-fundamentals/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
annotated-types==0.7.0
    # via
    #   -r ../requirements.txt
    #   pydantic
anyio==4.12.1
    # via
    #   -r ../requirements.txt
    #   httpx
authlib==1.6.9
    # via
    #   -r ../requirements.txt
    #   weaviate-client
certifi==2026.2.25
    # via
    #   -r ../requirements.txt
    #   httpcore
    #   httpx
cffi==2.0.0
    # via
    #   -r ../requirements.txt
    #   cryptography
cryptography==46.0.5
    # via
    #   -r ../requirements.txt
    #   authlib
deprecation==2.1.0
    # via
    #   -r ../requirements.txt
    #   weaviate-client
grpcio==1.78.0
    # via
    #   -r ../requirements.txt
    #   weaviate-client
h11==0.16.0
    # via
    #   -r ../requirements.txt
    #   httpcore
httpcore==1.0.9
    # via
    #   -r ../requirements.txt
    #   httpx
httpx==0.28.1
    # via
    #   -r ../requirements.txt
    #   weaviate-client
idna==3.11
    # via
    #   -r ../requirements.txt
    #   anyio
    #   httpx
numpy==2.4.3
    # via pandas
packaging==26.0
    # via
    #   -r ../requirements.txt
    #   deprecation
pandas==3.0.1
    # via -r requirements.in
protobuf==6.33.6
    # via
    #   -r ../requirements.txt
    #   weaviate-client
pycparser==3.0
    # via
    #   -r ../requirements.txt
    #   cffi
pydantic==2.12.5
    # via
    #   -r ../requirements.txt
    #   weaviate-client
pydantic-core==2.41.5
    # via
    #   -r ../requirements.txt
    #   pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.2
    # via -r ../requirements.txt
six==1.17.0
    # via python-dateutil
typing-extensions==4.15.0
    # via
    #   -r ../requirements.txt
    #   anyio
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   -r ../requirements.txt
    #   pydantic
tzdata==2025.3
    # via pandas
validators==0.35.0
    # via
    #   -r ../requirements.txt
    #   weaviate-client
weaviate-client==4.20.4
    # via
    #   -r ../requirements.txt
    #   -r requirements.in
